@@ -14,9 +14,11 @@
     "@vuelidate/validators": "^2.0.0",
     "c3": "^0.7.20",
     "model": "https://github.com/twhtanghk/model",
+    "node-polyfill-webpack-plugin": "^2.0.1",
     "node-sass": "^8.0.0",
     "register-service-worker": "^1.6.2",
-    "vue": "^2.7.13"
+    "vue": "^3.2.47",
+    "vue-cli-plugin-coffee": "^1.0.2"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^5.0.8",
@@ -34,10 +36,10 @@
     "mocha": "^10.1.0",
     "sass-loader": "^13.2.0",
     "supertest": "^6.3.1",
-    "vue-router": "^3.6.5",
+    "vue-router": "^4.1.6",
     "vue-template-compiler": "^2.6.11",
     "vue.model": "https://github.com/twhtanghk/vue.model",
-    "vuetify": "^2.1.15"
+    "vuetify": "^3.1.10"
   },
   "eslintConfig": {
     "root": true,
